apartment-matcher/requirements.txt
beautifulsoup4==4.12.2
selenium==4.11.2
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
pandas==2.1.3
pydantic==2.5.1
schedule==1.2.0
python-multipart==0.0.6